[Server]: Fix on port check in cli args. [Client]: Now updates the spreadsheet name box when a user clicks from the list of spreadsheets.

+++ server/Utilities/AParser.cpp
@@ -85,7 +85,7 @@ namespace RS {
 	 *
 	 * @return JSON object of this class.
 	 */
-	uint16_t AParser::Port() {
+	int AParser::Port() {
 		return port_;
 	}
